feat: secure OpenAI API key with server-side route
Move AI analysis to server-side to protect API key and add auth/authz.

#VERCEL_SKIP
